backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

sqlalchemy==2.0.30
alembic==1.13.1
asyncpg
psycopg2-binary==2.9.9

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==1.2.0

openai==1.30.1
google-generativeai==0.5.4

scikit-learn
pandas
numpy
joblib

pypdf2==3.0.1
python-docx==1.1.0
pillow
pytesseract==0.3.10

pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.24.1
boto3==1.34.100

pytest==8.2.0
pytest-asyncio==0.23.6